+++ app/src/test/java/bc10/Selenium/Test1.java
@@ -1,7 +1,6 @@
 package bc10.Selenium;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -10,33 +9,18 @@ import org.openqa.selenium.Dimension;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.support.PageFactory;
-import org.openqa.selenium.support.ui.Select;
 
 import java.time.Duration;
 
-public class ComputerDataBase {
-
+public class Test1 {
     WebDriver driver;
 
     By btnFilterNameloc = By.xpath("//input[@id='searchsubmit']");
-
     By btnAddNewComputerloc = By.xpath("//a[@id='add']");
-
     By btnHomeloc = By.xpath("//a[normalize-space()='Computer database']");
-
     By btnComputerNameloc = By.xpath("//a[normalize-space()='Computer database']");
     By btnDiscontinuedloc = By.xpath("//a[normalize-space()='Discontinued']");
     By btnCompanyloc = By.xpath("//a[normalize-space()='Company']");
-    By btnCreateThisComputerloc = By.xpath("//input[@value='Create this computer']");
-    By mensajeloc = By.xpath("//span[contains(text(),'Failed to refine type : Predicate isEmpty() did no')]");
-
-    By searchboxComputerNameloc = By.xpath("//input[@id='name']");
-    By searchboxIntroducedloc = By.xpath("//input[@id='introduced']");
-    By searchboxDiscontinuedloc = By.xpath("//input[@id='discontinued']");
-    By selectCompanyloc = By.xpath("//select[@id='company']");
-    By msjeloc = By.xpath("//div[@class='alert-message warning']");
-
 
     @BeforeEach
     void setUp() {
@@ -46,11 +30,11 @@ public class ComputerDataBase {
         driver.get("https://computer-database.gatling.io");
         driver.manage().deleteAllCookies();
         driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(3));
+
     }
 
-
     @Test
-    void Test1() throws InterruptedException {
+    void Test1(){
 
         driver.manage().window().setSize(new Dimension(800,600));
 
@@ -59,7 +43,6 @@ public class ComputerDataBase {
         int widht = dimension.getWidth();
         int heigth = dimension.getHeight();
 
-        Thread.sleep(2000);
         // valida las dimensiones
         Assertions.assertEquals(800,widht);
         Assertions.assertEquals(600,heigth);
@@ -74,12 +57,13 @@ public class ComputerDataBase {
         //Boton nuevo computador
         WebElement btnAddNewComputer = driver.findElement(btnAddNewComputerloc);
         btnAddNewComputer.click();
+        //Entra a add computer y valida que esta correcta la entrada con el titulo de la pagina
         Assertions.assertEquals("Computers database",driver.getTitle());
 
         //Regresa a la home
         WebElement btnHome = driver.findElement(btnHomeloc);
         btnHome.click();
-
+        // validacion de que nos envia a algun destino.
         //Computer name
         WebElement btnComputerName = driver.findElement(btnComputerNameloc);
         btnComputerName.click();
@@ -94,64 +78,4 @@ public class ComputerDataBase {
         Assertions.assertEquals("Computers database",driver.getTitle());
 
     }
-
-    @Test
-    void test2(){
-
-        WebElement btnAddNewComputer = driver.findElement(btnAddNewComputerloc);
-        btnAddNewComputer.click();
-
-        driver.manage().window().maximize();
-
-        WebElement btnCreateNewComputer = driver.findElement(btnCreateThisComputerloc);
-        btnCreateNewComputer.click();
-
-        WebElement mensaje = driver.findElement(mensajeloc);
-
-        Assertions.assertEquals("Failed to refine type : Predicate isEmpty() did not fail.",mensaje.getText());
-
-
-    }
-
-    @Test
-    void Test3(){
-
-        WebElement btnAddNewComputer = driver.findElement(btnAddNewComputerloc);
-        btnAddNewComputer.click();
-        driver.manage().window().maximize();
-
-        WebElement searchboxCompuerName = driver.findElement(searchboxComputerNameloc);
-        searchboxCompuerName.sendKeys("BC10_JJMN");
-
-        WebElement searchboxIntroduced = driver.findElement(searchboxIntroducedloc);
-        searchboxIntroduced.sendKeys("2022-10-23");
-
-        WebElement searchboxDiscontinued = driver.findElement(searchboxDiscontinuedloc);
-        searchboxDiscontinued.sendKeys("2022-10-24");
-
-        WebElement selectCompany = driver.findElement(selectCompanyloc);
-        selectCompany.click();
-
-        Select company = new Select(selectCompany);
-        company.selectByIndex(42);
-
-
-        WebElement btnCreateNewComputer = driver.findElement(btnCreateThisComputerloc);
-        btnCreateNewComputer.click();
-
-        WebElement msje = driver.findElement(msjeloc);
-
-
-        Assertions.assertEquals("Done ! Computer BC10_JJMN has been created",msje.getText());
-
-    }
-
-
-
-    @AfterEach
-    void quit(){
-        driver.quit();
-    }
-
 }
-
